🎨 PromiseSpoolerを分離
