+++ src-electron/util/promiseSpooler.ts
@@ -1,0 +1,74 @@
+/** 複数の処理を並列で受け取って直列で処理する */
+export class PromiseSpooler {
+    /** 待機中の処理のQueue */
+    spoolingQueue: [
+        () => Promise<any>,
+        (value: any | PromiseLike<any>) => void,
+        undefined | string
+    ][];
+    running: boolean;
+
+    constructor() {
+        this.spoolingQueue = [];
+        this.running = false;
+    }
+
+    async pushItem<T>(
+        spoolingQueue: [
+            () => Promise<any>,
+            (value: any) => void,
+            string | undefined
+        ][],
+        process: () => Promise<T>,
+        resolve: (value: T | PromiseLike<T>) => void,
+        channel: string | undefined
+    ) {
+        if (channel !== undefined) {
+            const lastItem = spoolingQueue[spoolingQueue.length - 1];
+            if (lastItem !== undefined) {
+                const [, lastResolve, lastChannel] = lastItem;
+                if (lastChannel === channel) {
+                    const newResolve = (value: T) => {
+                        lastResolve(value);
+                        resolve(value);
+                    };
+                    // チャンネルが同じ場合は処理を上書き
+                    // resolveは統合
+                    lastItem[0] = process;
+                    lastItem[1] = newResolve;
+                    return;
+                }
+            }
+        }
+        // それ以外の場合処理を追加
+        spoolingQueue.push([process, resolve, channel]);
+    }
+
+    /** channelを指定すると同じchannelの処理は連続せず上書きされる */
+    async spool<T>(spollingItem: () => Promise<T>, channel?: string) {
+        const pushItem = (
+            process: () => Promise<T>,
+            resolve: (value: T | PromiseLike<T>) => void,
+            channel: string | undefined
+        ) => this.pushItem(this.spoolingQueue, process, resolve, channel);
+
+        const resultPromise = new Promise<T>((resolve) => {
+            pushItem(spollingItem, resolve, channel);
+        });
+        this.start();
+        return resultPromise;
+    }
+
+    private async start() {
+        if (this.running) return;
+        this.running = true;
+        while (true) {
+            const item = this.spoolingQueue.shift();
+            if (item === undefined) break;
+            const [process, resolve] = item;
+            const result = await process();
+            resolve(result);
+        }
+        this.running = false;
+    }
+}
